api/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests
google-auth